Change split ratio fields from int to float
Updated the `numerator` and `denominator` fields in the `SplitEvent` model to use floats instead of integers. This ensures support for split ratios with

+++ fmp_data/intelligence/models.py
@@ -92,8 +92,8 @@ class StockSplitEvent(BaseModel):
     symbol: str = Field(description="Company symbol")
     split_event_date: date = Field(description="Split date", alias="date")
     label: str = Field(description="Human-readable date label")
-    numerator: int = Field(description="Numerator of the split ratio")
-    denominator: int = Field(description="Denominator of the split ratio")
+    numerator: float = Field(description="Numerator of the split ratio")
+    denominator: float = Field(description="Denominator of the split ratio")
 
 
 class IPOEvent(BaseModel):
